# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Logbook PPDS API Specification
  description: API Documentation & Testing Interface for Logbook PPDS Backend (Golang & Gin).
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Server
paths:
  /api/v1/auth/login:
    post:
      summary: User Login
      description: Authenticate user using username and password (and optional role).
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: residen01
                password:
                  type: string
                  example: residen01
      responses:
        "200":
          description: Login Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginSuccessResponse"
        "401":
          description: Invalid Credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/auth/me:
    get:
      summary: Get Logged-in User Profile
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Profile Details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardResponse"
        "401":
          description: Unauthorized
  /api/v1/users:
    get:
      summary: List All Users (Super Admin Only)
      tags:
        - User Management
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardResponse"
        "403":
          description: Forbidden - Insufficient Role

    post:
      summary: Create New User (Super Admin Only)
      tags:
        - User Management
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - name
                - email
                - password
                - role
              properties:
                username:
                  type: string
                  example: dokter02
                name:
                  type: string
                  example: dr. Ahmad Fauzi
                email:
                  type: string
                  example: ahmad@ppds.id
                password:
                  type: string
                  example: password123
                role:
                  type: string
                  enum: [residen, supervisor, admin, superadmin]
                  example: residen
                jabatan:
                  type: string
                  example: Residen Bedah Th.1
      responses:
        "201":
          description: User Created Successfully
  /api/v1/users/{id}:
    get:
      summary: Get User by ID (Super Admin Only)
      tags:
        - User Management
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: User Details
        "404":
          description: User Not Found

    put:
      summary: Update User (Super Admin Only)
      tags:
        - User Management
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - role
              properties:
                name:
                  type: string
                  example: dr. Ahmad Fauzi, Sp.B
                email:
                  type: string
                  example: ahmad.new@ppds.id
                password:
                  type: string
                  example: newpassword123
                role:
                  type: string
                  example: supervisor
                jabatan:
                  type: string
                  example: DPJP Bedah Umum
      responses:
        "200":
          description: User Updated Successfully

    delete:
      summary: Delete User (Super Admin Only)
      tags:
        - User Management
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 2
      responses:
        "200":
          description: User Deleted Successfully
  /api/v1/jadwals:
    get:
      summary: Get FullCalendar Schedule Events
      description: Get schedule events filtered by start, end date, or type (jaga, bimbingan, rotasi, ujian).
      tags:
        - Jadwal Management
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: start
          schema:
            type: string
          example: "2026-08-01"
        - in: query
          name: end
          schema:
            type: string
          example: "2026-08-31"
        - in: query
          name: type
          schema:
            type: string
            enum: [all, jaga, bimbingan, rotasi, ujian]
          example: all
      responses:
        "200":
          description: Array of FullCalendar Events

    post:
      summary: Create Schedule Event
      tags:
        - Jadwal Management
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - start
                - end
                - type
              properties:
                title:
                  type: string
                  example: Jaga Malam IGD Shift 2
                description:
                  type: string
                  example: Menjaga pasien IGD bersama tim 2
                location:
                  type: string
                  example: Ruang IGD RSUD
                start:
                  type: string
                  example: "2026-08-25 20:00:00"
                end:
                  type: string
                  example: "2026-08-26 08:00:00"
                all_day:
                  type: boolean
                  example: false
                type:
                  type: string
                  enum: [jaga, bimbingan, rotasi, ujian]
                  example: jaga
      responses:
        "200":
          description: Event Created Successfully
  /api/v1/jadwals/{id}:
    put:
      summary: Update Schedule Event
      tags:
        - Jadwal Management
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - start
                - end
                - type
              properties:
                title:
                  type: string
                  example: Jaga Malam IGD Shift 1 Update
                description:
                  type: string
                  example: Deskripsi perbaikan
                location:
                  type: string
                  example: IGD Gedung Utama
                start:
                  type: string
                  example: "2026-08-25 20:00:00"
                end:
                  type: string
                  example: "2026-08-26 08:00:00"
                all_day:
                  type: boolean
                  example: false
                type:
                  type: string
                  example: jaga
      responses:
        "200":
          description: Event Updated Successfully

    delete:
      summary: Delete Schedule Event
      tags:
        - Jadwal Management
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Event Deleted Successfully
  /api/v1/jadwals/{id}/dates:
    patch:
      summary: Update Schedule Dates for Drag & Drop
      tags:
        - Jadwal Management
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - start
                - end
              properties:
                start:
                  type: string
                  example: "2026-08-26 20:00:00"
                end:
                  type: string
                  example: "2026-08-27 08:00:00"
                all_day:
                  type: boolean
                  example: false
      responses:
        "200":
          description: Event Dates Updated Successfully
  /api/v1/tindakans:
    get:
      summary: Get Logbook Summary & Tindakan Entries
      description: Get rekap statistik logbook (total, mandiri, dibimbing, observasi, disetujui, percent) and list of procedure entries.
      tags:
        - Tindakan Logbook
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Logbook Summary & Entries List

    post:
      summary: Create New Tindakan (Draft)
      tags:
        - Tindakan Logbook
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - mr_number
                - patient_name
                - diagnosis_label
                - plan_procedure
              properties:
                mr_number:
                  type: string
                  example: RM-88406
                visit_date:
                  type: string
                  example: "2026-08-07"
                patient_name:
                  type: string
                  example: Tn. Ahmad Hidayat
                gender:
                  type: string
                  enum: [L, P]
                  example: L
                birth_date:
                  type: string
                  example: "1992-05-15"
                division:
                  type: string
                  example: Trauma Muskoloskeletal
                diagnosis_label:
                  type: string
                  example: Fraktur Femur Dekstra
                procedure_code:
                  type: string
                  example: ORIF-PLT-01
                plan_procedure:
                  type: string
                  example: ORIF Plating Femur
                activity:
                  type: string
                  example: Bedah Sentral (OK)
                procedure_date:
                  type: string
                  example: "2026-08-07"
                room_label:
                  type: string
                  example: OK 2 Central
                role_label:
                  type: string
                  example: Operator Utama
                kemandirian:
                  type: string
                  enum: [mandiri, dibimbing, observasi]
                  example: mandiri
                clinical_note:
                  type: string
                  example: Reduksi anatomis sukses
                supervisor_name:
                  type: string
                  example: dr. Andi Wijaya, Sp.OT
      responses:
        "201":
          description: Tindakan Draft Created Successfully
  /api/v1/tindakans/getdpjp:
    get:
      summary: Get DPJP List
      description: Get list of supervisor names from users table (role = supervisor) for DPJP dropdown.
      tags:
        - Tindakan Logbook
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of DPJP names
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: Berhasil mengambil daftar DPJP
                  data:
                    type: array
                    items:
                      type: string
                    example:
                      - dr. Andi Wijaya, Sp.OT
                      - dr. Siti Rahma, Sp.OT
        "401":
          description: Unauthorized
  /api/v1/tindakans/{id}:
    get:
      summary: Get Single Tindakan Detail
      tags:
        - Tindakan Logbook
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Tindakan Detail

    put:
      summary: Update Tindakan
      tags:
        - Tindakan Logbook
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - mr_number
                - patient_name
                - plan_procedure
              properties:
                mr_number:
                  type: string
                  example: RM-88406
                patient_name:
                  type: string
                  example: Tn. Ahmad Hidayat
                plan_procedure:
                  type: string
                  example: ORIF Plating Femur Update
                diagnosis_label:
                  type: string
                  example: Fraktur Femur Update
                room_label:
                  type: string
                  example: OK 1
                clinical_note:
                  type: string
                  example: Catatan perbaikan
      responses:
        "200":
          description: Tindakan Updated Successfully

    delete:
      summary: Delete Tindakan
      tags:
        - Tindakan Logbook
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Tindakan Deleted Successfully
  /api/v1/tindakans/{id}/send:
    post:
      summary: Send Tindakan to DPJP Approval (draft -> menunggu)
      tags:
        - Tindakan Logbook
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 4
      responses:
        "200":
          description: Sent to DPJP for Verification Successfully
  /api/v1/kegiatan-ilmiah:
    get:
      summary: Get Kegiatan Ilmiah (Grouped by Kategori)
      description: Get all kegiatan ilmiah entries grouped by kategori (simposium, workshop, multidisiplin, ilmiah_lain)
      tags:
        - Kegiatan Ilmiah
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Kegiatan Ilmiah Grouped by Kategori
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object

    post:
      summary: Create New Kegiatan Ilmiah
      description: Create kegiatan ilmiah. user_username, program_studi, ppds_name, nim_nip diambil dari session user yang login.
      tags:
        - Kegiatan Ilmiah
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - kategori
                - jenis_kegiatan
                - topik
              properties:
                kategori:
                  type: string
                  enum: [simposium, workshop, multidisiplin, ilmiah_lain]
                  example: simposium
                jenis_kegiatan:
                  type: string
                  example: Seminar Nasional Ilmiah
                topik:
                  type: string
                  example: Perkembangan Terbaru Oftalmologi
                tanggal_mulai:
                  type: string
                  example: "2026-08-15"
                tanggal_selesai:
                  type: string
                  example: "2026-08-16"
                lokasi_tipe:
                  type: string
                  enum: [rsds_fk_unair, luar_rsds_fk_unair]
                  example: rsds_fk_unair
                lokasi_detail:
                  type: string
                  example: Ruang Auditorium RSUD dr. Soebandi
                sebagai:
                  type: string
                  example: Peserta / Pembicara
                pembimbing_1:
                  type: string
                  example: dr. Bambang Prasetyo, Sp.M
                pembimbing_2:
                  type: string
                  example: dr. Siti Rahma, Sp.M
                pembimbing_3:
                  type: string
                pembimbing_4:
                  type: string
                pembimbing_5:
                  type: string
                penguji_1:
                  type: string
                penguji_2:
                  type: string
                penguji_3:
                  type: string
                penguji_4:
                  type: string
                penguji_5:
                  type: string
                deskripsi:
                  type: string
                  example: Deskripsi kegiatan ilmiah
      responses:
        "201":
          description: Kegiatan Ilmiah Created Successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  kegiatan:
                    type: object
                    properties:
                      id:
                        type: integer
                      user_username:
                        type: string
                      program_studi:
                        type: string
                      ppds_name:
                        type: string
                      nim_nip:
                        type: string
                      kategori:
                        type: string
                      jenis_kegiatan:
                        type: string
                      topik:
                        type: string
                      status:
                        type: string
                        enum: [pending, disetujui, ditolak]
  /api/v1/kegiatan-ilmiah/{id}:
    delete:
      summary: Delete Kegiatan Ilmiah
      tags:
        - Kegiatan Ilmiah
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Kegiatan Ilmiah Deleted Successfully
  /api/v1/bimbingans:
    get:
      summary: Get All Bimbingan Penelitian
      description: Get list of all bimbingan penelitian sessions
      tags:
        - Bimbingan Penelitian
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of Bimbingan Penelitian

    post:
      summary: Create New Bimbingan Penelitian
      description: Create bimbingan penelitian. user_username diambil dari session user yang login.
      tags:
        - Bimbingan Penelitian
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tahap
                - topik_bimbingan
                - pembimbing
                - catatan_pembimbing
              properties:
                tahap:
                  type: string
                  example: Proposal / Metodologi
                topik_bimbingan:
                  type: string
                  example: Pemilihan topik penelitian dan metodologi
                pembimbing:
                  type: string
                  example: dr. Hendra Wijaya, M.Kes
                catatan_pembimbing:
                  type: string
                  example: Topik sudah sesuai, lanjutkan ke tahap berikutnya
      responses:
        "201":
          description: Bimbingan Penelitian Created Successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  bimbingan:
                    type: object
                    properties:
                      id:
                        type: integer
                      user_username:
                        type: string
                      sesi_ke:
                        type: string
                      tahap:
                        type: string
                      topik_bimbingan:
                        type: string
                      pembimbing:
                        type: string
                      catatan_pembimbing:
                        type: string
                      status_acc:
                        type: string
                      created_at:
                        type: string
                      updated_at:
                        type: string
...
  /api/v1/pendidikan/rotasi:
    get:
      summary: Get All Pendidikan Rotasi
      tags:
        - Pendidikan
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of Rotasi entries
    post:
      summary: Create Rotasi Entry
      description: Create rotasi. user_username diambil dari session.
      tags:
        - Pendidikan
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shift:
                  type: string
                  example: Dinas Pagi
                pembimbing:
                  type: string
                  example: dr. Andi Wijaya, Sp.OT
                catatan:
                  type: string
                  example: Presensi harian
      responses:
        '201':
          description: Rotasi created successfully
...
  /api/v1/approval/disetujui:
    get:
      summary: Get Approved Items
      description: Get list of approved items - Tindakan, Kegiatan Ilmiah, Aktivitas Klinik, Pendidikan Evaluasi
      tags:
        - Approval
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of approved items
  /api/v1/approval/ditolak:
    get:
      summary: Get Rejected Items
      description: Get list of rejected items needing revision
      tags:
        - Approval
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of rejected items
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginSuccessResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: Login Success
        data:
          type: object
          properties:
            username:
              type: string
              example: residen01
            name:
              type: string
              example: dr. Ratna Puspita
            role:
              type: string
              example: residen
            jabatan:
              type: string
              example: Residen Bedah Th.2
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6...
    StandardResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: Success
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 401
        message:
          type: string
          example: Username, kata sandi, atau peran yang dipilih tidak cocok.
...